==> Reports/01_readonly.rmd <==
---
title: "R Notebook"
output: html_notebook
---

#1. Trim data (tGPS())
#2. Group GPS and add falcon data
#3. Finding optimal site radius
#4. Trim data by defined site radius


# Trim data (tGPS())
```{R}

# libraries
library(stringr)
library(robofalcon)

# files / folders 
fold  = file.path( PROJHOME , "Data" , "pigeon")
files = list.files(fold)
folderDS (c("Output", "tGPS"))

#loop
for ( i in 1:length(files)){
  #i=155
  load( file.path( fold, files[i]))
  dat = data
  data = tGPS( dat  , plot_traj =F)
  save( data, file = file.path(PROJHOME , "Output" , "tGPS" , files[i]))
  print( paste( i , "/" , length(files)))
}
```

# Group GPS and add falcon data

```{R}
{ # skip to loop 
  
  #housekeeping
  rm(list=ls())
  
  #objects 
  plot_traj = T
  plot_speed = F
  plot2pdf = F
  pfs = F # preflightsection
  site.radius = 700
  
  #libraries
  library(stringr)
  library(robofalcon)
  
  #files/folders
  folder.f = file.path( PROJHOME , "Data" , "Falcon")
  folder.p  = file.path( PROJHOME , "Output" , "tGPS")
  files.p = list.files(folder.p)
  files.f = list.files(folder.f)
  folderDS ( c("Output", "gGPS", paste0("site-radius",site.radius) ))
  folderDS ( c("Output", "gGPS", "pre-flight"))
  if( pfs){
    folder.out = file.path( PROJHOME , "Output" , "gGPS", "pre-flight")
  } else {
    folder.out = file.path(PROJHOME , "Output", "gGPS", paste0("site-radius", site.radius))
  }
  #flight variables
  mat = str_split_fixed(files.p, "\\.",12)
  flight = as.numeric( mat[,1])
  day = as.numeric(mat[!duplicated(flight),7])
  predation = mat[!duplicated(flight),3]
  falcon.day = as.numeric(substr(files.f,8,9))
  
  #figure 
  if( plot2pdf){
    pdf ( file.path ( PROJHOME , "Figures" , "Trajectories.pdf"))
  }
  
}
# loop
for ( i in 1:length(unique(flight))){
  # i=19
  
  # group pigeons 
  {
    files2 = files.p[flight == i]
    data.p = gGPS(files2, folder.p,plot_traj = plot_traj,plot_speed = plot_speed, pre.flight.section = pfs, site.radius = site.radius)
    pidge.num = dim(data.p)[3]
  }
  
  # # # work
  # key.code = register_ggmap(install = F)
  # plot_map(data.p, key.code, register = F, type = "group" ,zoom =16)
  # # # 
  
  # file name
  vec= mat[flight==i,][1,c(3,1,4,5,6)]
  file.name = paste0(c(vec[2:1],"g",vec[3],"f",vec[4],vec[5], "rda"),collapse = ".")
  
  #if falcon flight - add falcon 
  if( predation[i] == "p"){
    load( file.path( folder.f , files.f[which(falcon.day == day[i])]))
    start = which( data[,3] == data.p[1,3,1])
    pred = data[start:(start-1+dim(data.p)[1] ),]
    data = data.p
    data.list = list(data,pred)
    
    ## PLOT ##
    lim = xylims( data)
    if ( plot_traj){
      {plot(data[,"lon",1],data[,"lat",1],type = "n", ylim = lim[[2]], xlim = lim[[1]])
        for ( k in 1:(pidge.num) ){
          lines ( data[,"lon",k],data[,"lat",k])
        }
        lines ( pred[,"lon"],pred[,"lat"] , col = 2)
      }
    }
    #save
    save( data.list, file = file.path(folder.out,file.name ))
    
  } else { 
    
    # otherwise no falcon
    data = data.p
    
    ## PLOT ## 
    if ( plot_traj){
      lim = xylims(data)
      if( predation[i] != "t"){
        {plot(data.p[,"lon",1],data.p[,"lat",1],type = "n", ylim = lim[[2]], xlim = lim[[1]])
          for ( k in 1:pidge.num){
            lines ( data.p[,"lon",k],data.p[,"lat",k], col = k)
          }
        }
      }
    }
    
    #save
    save( data, file = file.path(folder.out,file.name ))
  }
  
  #take stock
  print(paste(i, "/", length(unique(flight))))
}

# Figures
if( plot2pdf){
  dev.off()
}

```

# What are the coordinates of the release site location? 

```{R}
#housekeeping
rm(list=ls())

# libraries
library(robofalcon)

# files / folders
fold = file.path(PROJHOME , "Output", "gGPS", "site-radius700")
files = list.files(fold)
spl = str_split_fixed(files, "\\.",12)

#empty objects
site.estimates = matrix(NA,0,2)

#loop 
for ( i in 1:length(files)){
  #i=1
  
  # data
  load( file.path(fold, files[i]))
  
  # add predator if "p"
  if( spl[i,2] == "p"){
    data = data.list[[1]]
  }
  
  # cent
  cent = d2c(data,return_centorid = T)
  site = cent[1,]
  
  # save
  site.estimates = rbind(site.estimates,site)
}

site.xy = c( lon = median(site.estimates[9:nrow(site.estimates),1]),
             lat = median(site.estimates[9:nrow(site.estimates),2]))

save (site.xy , file = file.path(PROJHOME , "metadata", "site-xy.rda"))

list ( cc = site.xy ,
       training =  c( lon = median(site.estimates[1:8,1]),
                      lat = median(site.estimates[1:8,2])))
```

# Finding optimal site radius

```{R}
{ # skip to loop
  
  # housekeeping
  rm( list = ls())
  
  # libraries
  library(robofalcon)
  library(stringr)
  
  # files / folders
  fold = file.path( PROJHOME , "Output", "gGPS" , "site-radius700")
  files = list.files(fold)
  spl = str_split_fixed(files, "\\.",12)
  
  # only predator files
  pred.files = which( spl[,2] == "p")
  
  # metadata
  load( file.path(PROJHOME ,"metadata" , "site-xy.rda"))
  
  # empty objects
  predxy = matrix(NA, 0,3)
  p2ps = matrix(NA,0,2)
  
}
# loop
for ( i in pred.files){
  #i=pred.files[1]
  
  # data
  load(file.path(fold, files[i]))
  
  # objects
  data = data.list[[1]]
  pred = data.list[[2]]
  len = dim(data)[[3]]
  
  # centroid 
  cent = d2c(data,return_centorid = T)
  
  # dist 2 pred
  d2p = get_dist(cent[,"lon"],cent[,"lat"],
                 pred[,"lon"],pred[,"lat"],method = "distance")
  
  # dist 2 site
  cent2site = get_dist(site.xy["lon"], site.xy["lat"],
                       cent[,1],cent[,2],method = "distance")
  pred2site = get_dist(site.xy["lon"], site.xy["lat"],
                       pred[,1], pred[,2], method = "distance")
  
  # for the relationship between predator's distance to site and distance to group
  pred2pidgesite = cbind(d2p = d2p, pred2site = pred2site)
  
  # save pred
  predxy = rbind(predxy,pred)
  p2ps = rbind(p2ps,pred2pidgesite)
  
  
}


# PLOTS
# 
# # Relationship between predator's distance to site and distance to group
# p2ps = as.data.frame(p2ps)
# library(ggplot2)
# g1 = ggplot( p2ps, aes ( y = d2p , x = pred2site))+
#   geom_point()+
#   xlab ("Distance from predator to release site (m)")+
#   ylab ("Distance from predator to flock centroid (m)")
# 
# # Predator trajectory on the map
# key.code = register_ggmap(install = T)
# g2 = plot_map(predxy, type = "solo" ,zoom =16, centre = site.xy, lines.or.points = "points")
# 
# gridExtra::grid.arrange(g1,g2,ncol = 2)
```


# Trim data by defined site radius, identify fission, and subgroups analysis

```{R}

{ # skip to loop
  
  # housekeeping
  rm(list= ls())
  
  # libraries 
  library(stringr)
  library(robofalcon)
  
  #objects
  split.dist = 10
  
  #files/folders 
  fold = file.path(PROJHOME , "Output", "gGPS", "site-radius700")
  files = list.files(fold)
  
  # splitstring
  spl = str_split_fixed(files, "\\.", 12)
  
  # empty objects 
  maxdists = rep(NA,length(files))
  subgroups = matrix( NA, 0,4,dimnames = list(NULL, c("subgroupsize" , "groupsize", "flight" , "timestep")))
}

# loop
for ( i in 1:
      length(files)){
  # 14){
  #i=17
  
  # data
  {
    load(file.path(fold, files[i]))
    if( spl[i,2] == "p" ){
      dat = data.list[[1]]
    } else {
      dat = data
    }
    dims = dim(dat)
  }
  
  # empty objects
  {
    # objects for if( d2site < 500) below
    in.flock.matr = matrix(NA, dims[1],dims[3])
    d2site = 0
    d1 = t(dat[1,1:2,])
    site1 = c( lon = mean(d1[,"lon"],na.rm=T),
               lat = mean(d1[,"lat"],na.rm=T))
    maxdist = vector()
  }
  
  # loop
  for ( j in 1:dims[1]){
    #j=77
    
    if( d2site < 500){ #see methods for 500 justification
      
      # deal with missing data
      dat.all = t(dat[j,1:2,])
      dat.all.complete = dat.all[complete.cases(dat.all),]
      dims.comp = dim(dat.all.complete)
      nam.all      = dimnames(dat.all)[[1]]
      nam.complete = dimnames(dat.all.complete)[[1]]
      whi = which ( nam.all %in% nam.complete)
      
      # convert to metres for groupDetect
      dat.metres = cbind( x = get_dist( dat.all.complete[,"lon"],
                                        site1["lat"],
                                        site1["lon"],
                                        site1["lat"],method = "distance"),
                          y = get_dist( site1["lon"],
                                        dat.all.complete[,"lat"],
                                        site1["lon"],
                                        site1["lat"],method = "distance"))
      
      # detect membership
      dist.mat = distance(dat.metres)
      group.mat = matrix(as.numeric(dist.mat<split.dist),nrow(dist.mat))
      group.vec = na.omit(as.vector( group.mat))
      if ( any(group.vec == 1)){
        # use Igraph to group individuals
        diag(group.mat) = 0
        net <- igraph::graph.adjacency(group.mat, mode='undirected',diag=FALSE)
        weights <- igraph::E(net)$weight
        coms <- suppressWarnings( igraph::edge.betweenness.community(net,weights=weights))
        membership = coms$membership
      } else {
        membership = 1:nrow(dat.metres)
      }
      
      # manipulate and add to matrix
      tb = table(membership)
      mainflock = which.max(tb)
      in.flock = membership == mainflock
      incl.missing      = rep(NA,dims[3]) # add back in any missing individuals (i.e. NA's) 
      incl.missing[whi] = in.flock
      in.flock.matr[j,] = incl.missing
      
      d2site = get_dist( mean(dat.all.complete[in.flock,"lon"]),
                         mean(dat.all.complete[in.flock,"lat"]),
                         site1["lon"],site1["lat"],method = "distance")
      
      # takestock
      k = j
      
      # save maxdist
      whi.memb = which(membership == mainflock)
      foo =  max(na.omit(as.vector(dist.mat[whi.memb,whi.memb]))) /dims[3]
      maxdist = c(maxdist,foo)
      
      
      
      #plot
      # XY = xylims(dat.metres)
      # 
      # plot( dat.metres[,1],dat.metres[,2],col = membership, ylim = XY[[2]],xlim = XY[[1]])
      # for ( i in 1:nrow(dat.metres)){
      #   x = dat.metres[i,1]
      #   y = dat.metres[i,2]
      #   draw.circle(x,y,split.dist,border =membership[i])
      # }
      
      
      # save subgroup analysis
      if ( length( tb) > 1){
        cb = cbind( tb[-mainflock],spl[i,7],i,j)
        subgroups= rbind(subgroups,cb)
        if ( spl[i,7] == "s" & any(as.numeric( tb[-mainflock]) > 5) ){
          stop()
        }
      }
      
    }
  }
  
  
  #save maxdists
  foo = max(maxdist)
  maxdists[i] = foo
  
  data.mainflock = dat[1:k,,]
  for ( j in 1:k){
    # j = 1
    in.f = in.flock.matr[j,]
    data.mainflock[j,,!in.f] = NA
  }
  data.entireflock = dat[1:k,,]
  data.list = list(data.mainflock, data.entireflock)
  
  
  # save
  folderDS ( c("Output", "DataRemovedOutliers"))
  save( data.list, file = file.path ( PROJHOME , "Output", "DataRemovedOutliers", files[i]))
  
  # takestock
  print( paste( i, "/" , length(files)))
}

# save subgroups
folderDS(c("Output","subgrouping"))
save( subgroups , file = file.path (PROJHOME , "Output", "subgrouping" , "subgroup.rda"))


```

## Trim data to site radius 500 

```{R}
# housekeeping
rm(list= ls())

#libraries
library(stringr)

# files/folders 
fold = file.path(PROJHOME , "Output", "gGPS", "site-radius700")
folderDS( c("Output", "gGPS", "site-radius500"))
fold.f = file.path(PROJHOME, "Output", "DataRemovedOutliers")
files = list.files(fold)

# splitstring
spl = str_split_fixed(files, "\\.", 12)

# loop
for (  i in 1:length(files)){
  #i=13
  
  # data from fission
  load( file.path(fold.f,files[i]))
  fis = data.list[[1]]
  k = dim(fis)[1]
  
  # data with all individuals and maybe falcon
  load( file.path(fold , files[i]))
  if( spl[i,2] == "p" ){
    dat = data.list[[1]]
    falc = data.list[[2]]
    data = dat[1:k,,]
    falc= falc[1:k,]
    data.list = list( data, falc)
    save( data.list, file = file.path (PROJHOME , "Output", "gGPS", "site-radius500", files[i]))
  } else {
    data = data[1:k,,]
    save( data     , file = file.path (PROJHOME , "Output", "gGPS", "site-radius500", files[i]))
    
  }
  
}
```

## How time lagged is the falcon?

```{R}
# housekeeping
rm(list= ls())

#libraries
library(stringr)

# files/folders 
fold = file.path(PROJHOME , "Output", "gGPS", "site-radius500")
files = list.files(fold)

# splitstring
spl = str_split_fixed(files, "\\.", 12)

# loop
par(mfrow = c(2,2))
for (  i in which ( spl[,2] == "p")){
  #i=9
  
  # data with all individuals and  falcon
  load( file.path(fold , files[i]))
  
  dat = data.list[[1]]
  falc = data.list[[2]]
  d2 = d2c(dat,return_centorid = T)
  mcent = lonlat2metres(d2[,1],d2[,2])
  mfalc = lonlat2metres(falc[,1],falc[,2])
  lRs = try(leadRship(mcent,mfalc, window=41,plot.cor = T))
  
  
}
```

# What is the autocorrelation of turning angles, speed, centr. acc,  

```{r}

{ # skip to loop
  
  # housekeeping
  rm(list = ls())
  
  # objects 
  plott = F
  
  # libraries
  library(robofalcon)
  
  # files / folders 
  fold = file.path(PROJHOME, "Output", "DataRemovedOutliers")
  files = list.files(fold)
  
  # empty objects
  aeps.turn = vector()
  aeps.speed= vector()
  aeps.centrip = vector()
  aeps.speedvar = vector()
  aeps.d2c = vector() 
  aeps.nn1 = vector()
}

# loop
for ( i in 9:length(files)){
  #i=9
  
  # data
  load( file.path(fold, files[i]))
  data = data.list[[2]]
  dims = dim(data)
  
  # obj for speedvar
  speed.mat= matrix(NA,(dims[1]),dims[3])
  
  for (j in 1:dims[3]){
    #j=1
    
    # focal heading and turning decisions.
    focal_head = get_heading(data[,"lon",j],data[,"lat",j])
    diff_head = turnang(data[,,j])
    alpha = 0.95
    ts =na.omit( diff_head)
    
    # speed
    speed1 = get_dist(data[,"lon",j],data[,"lat",j],hz = 5,method = "speed")
    speed.mat[,j] = speed1
    speed = na.omit(speed1)
    
    # centripetal
    turn.radius = speed1 / diff_head
    centrip = speed1^2 / turn.radius
    
    # autocorrelation
    ts.acf <- acf(na.omit(ts),plot = plott, lag.max = 50)
    conf.lim <- qnorm((1 + alpha)/2)/sqrt(ts.acf$n.used)
    aep = which(ts.acf$acf < conf.lim)[1]
    aeps.turn = na.omit(c(aeps.turn,aep))
    
    ts.acf <- acf(na.omit(speed),plot = plott, lag.max = 50)
    conf.lim <- qnorm((1 + alpha)/2)/sqrt(ts.acf$n.used)
    aep = which(ts.acf$acf < conf.lim)[1]
    aeps.speed = na.omit(c(aeps.speed,aep))
    
    ts.acf <- acf(na.omit(centrip),plot = plott, lag.max = 50)
    conf.lim <- qnorm((1 + alpha)/2)/sqrt(ts.acf$n.used)
    aep = which(ts.acf$acf < conf.lim)[1]
    aeps.centrip = na.omit(c(aeps.centrip,aep))
    
  }
  
  # speed variance
  speed.var = apply(speed.mat,1, function(x){
    x [ x<5 ]= NA
    x [ x>33]=NA
    sd(x,na.rm = T)/mean(x,na.rm = T)})
  ts.acf <- acf(na.omit(speed.var),plot = plott, lag.max = 50)
  conf.lim <- qnorm((1 + alpha)/2)/sqrt(ts.acf$n.used)
  aep = which(ts.acf$acf < conf.lim)[1]
  aeps.speedvar = na.omit(c(aeps.speedvar,aep))
  
  # flock cohesion
  dist2cent = d2c( data)
  nn  = near_nei(data)[[2]]
  d2c = na.omit(rowMeans( dist2cent , na.rm =  T))
  nn1 = na.omit(rowMeans( nn,na.rm=T))
  
  ts.acf <- acf(d2c,plot = plott, lag.max = 50)
  conf.lim <- qnorm((1 + alpha)/2)/sqrt(ts.acf$n.used)
  aep = which(ts.acf$acf < conf.lim)[1]
  aeps.d2c = na.omit(c(aeps.d2c,aep))
  
  ts.acf <- acf(nn1,plot = plott, lag.max = 50)
  conf.lim <- qnorm((1 + alpha)/2)/sqrt(ts.acf$n.used)
  aep = which(ts.acf$acf < conf.lim)[1]
  aeps.nn1 = na.omit(c(aeps.nn1,aep))
  
}

{
  #pdf( file.path (PROJHOME , "Figures", "Stats" , "Autocorrelation", "autocor.pdf"))
  
  par(mfrow= c(3,2))
  hist(aeps.turn/5,breaks = 50, xlim =c(0,10), main = "Turn angle", xlab = "Time (s)", col = "dark grey", cex.lab= 1.5)
  abline( v = median(aeps.turn)/5, lty = 2)
  list(shapiro.test(aeps.turn),
       median(aeps.turn),
       median(aeps.turn/5))
  
  hist(aeps.speed/5,breaks = 50,xlim =c(0,10), main = "Speed", xlab = "Time (s)", col = "dark grey", cex.lab= 1.5)
  abline( v = median(aeps.speed)/5, lty = 2)
  list(shapiro.test(aeps.speed),
       median(aeps.speed),
       median(aeps.speed/5))
  
  hist(aeps.speedvar/5,breaks = 50,xlim =c(0,10), main = "Speed variance", xlab = "Time (s)", col = "dark grey", cex.lab= 1.5)
  abline( v = median(aeps.speedvar)/5, lty = 2)
  list(shapiro.test(aeps.speedvar),
       median(aeps.speedvar),
       median(aeps.speedvar/5))
  
  hist(aeps.centrip/5,breaks = 50,xlim =c(0,10), main = "Centripetal acceleration", xlab = "Time (s)", col = "dark grey", cex.lab= 1.5)
  abline( v = median(aeps.centrip)/5, lty = 2)
  list(shapiro.test(aeps.centrip),
       median(aeps.centrip),
       median(aeps.centrip/5))
  
  hist(aeps.d2c/5,breaks = 50,xlim =c(0,10), main = "Distance to centroid", xlab = "Time (s)", col = "dark grey", cex.lab= 1.5)
  a= mean(aeps.d2c)/5
  b= 0.2  
  c= round(a/b)*b
  abline( v = c, lty = 2)
  list(shapiro.test(aeps.d2c),
       median(aeps.d2c),
       median(aeps.d2c/5),
       mean(aeps.d2c),
       mean(aeps.d2c/5))
  
  hist(aeps.nn1/5,breaks = 50,xlim =c(0,10), main = "Nearest neighbour distance", xlab = "Time (s)", col = "dark grey", cex.lab= 1.5)
  a= mean(aeps.nn1)/5
  b= 0.2  
  c= round(a/b)*b
  abline( v = c, lty = 2)
  list(shapiro.test(aeps.nn1),
       median(aeps.nn1),
       median(aeps.nn1/5),
       mean(aeps.nn1),
       mean(aeps.nn1/5))
  #dev.off()
}
```

## SPEED VARIANCE 

```{R}

{ # skip to loop
  
  # housekeeping
  rm(list = ls())
  
  # libraries
  library(robofalcon)
  library(stringr)
  library(lme4)
  library(nlme)
  
  
  # files / folders 
  fold = file.path(PROJHOME, "Output", "gGPS" , "site-radius500")
  files = list.files(fold)
  spl = str_split_fixed(files, "\\.",12)
  
  # empty objects 
  speed.vars = vector()
  ts = vector()
}

# loop
for ( i in 9:length(files)){ # not training flights
  #i=9
  # data
  load( file.path( fold, files[i]))
  
  # if predator
  if ( spl[i,2] == "p"){
    data = data.list[[1]]
  }
  
  # dims
  dims = dim(data)
  
  # speed 
  speeds = apply ( data , 3, function ( x){
    get_dist( x[,1],x[,2],hz = 5,method = "speed")
  })
  
  
  # speed variance
  speed.var = c( apply( speeds, 1,function(x){ 
    x [ x<5 ]= NA
    x [ x>33]=NA
    sd(x,na.rm = T)/mean(x,na.rm = T)
  }) , rep( NA,  dims[1]*(dims[3]-1) )) # object now length of total rows
  
  # add to dat
  speed.vars = c(speed.vars,speed.var)
  
  # takestock
  #print(paste (i , "/" , length(files)))
}






# Save
folderDS( c( "smaller-dataframes-for-git"))
save( speed.vars, file = file.path(PROJHOME, "smaller-dataframes-for-git" , "Speedvardat.rda"))
```

# plot_map? 

```{R}
if ( !file.exists(file.path( PROJHOME , "Figures", "fission_maps", "fission-maps.pdf"))){
  { # skip to loop
    
    # housekeeping
    rm(list = ls())
    
    # libraries
    library(robofalcon)
    
    # files/folders 
    fold = file.path ( PROJHOME , "Output", "DataRemovedOutliers")
    files = list.files(fold)
    
    # metadata
    load( file.path(PROJHOME ,"Data" ,"metadata" , "site-xy.rda"))
    
    # register
    register_ggmap(install = F)
    
  }
  
  {
    # pdf 
    pdf( file.path( PROJHOME , "Figures", "fission_maps", "fission-maps.pdf"))
    
    for ( i in 9:51){
      #i=9
      load( file.path(fold, files[i]))
      data.list = list(data.list[[2]],data.list[[1]])
      # p = robofalcon::plot_map( data.list,col = c("red","blue"), centre = site.xy,lines.or.points = "points", lwd = 1)
      
      d = data.list[[1]]
      
      plot ( d[,1,1], d[,2,1])
      print(p)
      print(i)
    }
    
    dev.off()
  }
}
```



# Big matrix of movement desisions, and potential covariates which predict it.

Run this script with only.remaining.indivs = T and = F

```{R}

{ # skip to loop
  
  # housekeeping
  rm(list = ls())
  
  # libraries
  library(robofalcon)
  library(stringr)
  library(plyr)
  library(circular)
  
  
  # objects
  only.small.groups = F
  only.remaining.indivs = T
  
  # files / folders 
  fold = file.path(PROJHOME, "Output", "gGPS" , "site-radius500")
  fold.remain = file.path(PROJHOME, "Output", "DataRemovedOutliers")
  files = list.files(fold)
  files.f=list.files(fold.remain)
  splitstring = str_split_fixed(files, "\\.",12)
  
  # metadata
  load( file.path( PROJHOME ,  "metadata" , "home-coordinates.rda"))
  load(file.path ( PROJHOME ,  "metadata" , "metadata-combinedPredLoft.rda"))
  
  #manip objects 
  splitstring = str_split_fixed(files, "\\." , 9)
  condition = splitstring[,2]
  splitstring = as.data.frame(splitstring)
  names(splitstring) = c("flight.id", as.character(2:ncol(splitstring)))
  if ( only.small.groups){
    whi.big = which(splitstring[,4] == 5)
    files = files[-whi.big]
    condition = condition[-whi.big]
    splitstring = splitstring[-whi.big,]
  }
  
  # empty objects
  aeps = vector()
  bigmatrix = matrix(NA,0,10)
  m = 1
  li = list()
}

# loop
for ( i in 9:length(files)){ # discounted earlier flights due to large differences in terrain, see Methods
  #i=9
  
  {
    # data
    load( file.path(fold, files[i]))
    if( condition[i] == "p"){
      # data = data.list[[1]]
      falc = data.list[[2]]
    } 
    load( file.path(fold.remain, files[i]))
    data.remain = data.list[[1]]
    data.all    = data.list[[2]]
    if( only.remaining.indivs){
      data= data.remain
    } else {
      data = data.all
    }
    
    dims = dim(data)
    
    # near neighbours 
    nn = near_nei(data)
    heads = apply( data, 3, function(x){
      get_heading(x[,"lon"],x[,"lat"])
    })
    heads.remain = apply( data.remain, 3, function(x){
      get_heading(x[,"lon"],x[,"lat"])
    })
    
    # centroid
    centroid = d2c(data,return_centorid = T)
    d2cent = d2c(data)
    
    # metadata
    md   =merge ( splitstring[i,], meta, by = "flight.id")
  }
  
  for (j in 1:dims[3]){
    #j=9
    {
    # metadata
    temp.data   =md[rep(1, dims[1]),c('condition','small.big','ind.flight','wind.direct.site', 'group.num','wind.speed.site','flight.time','Date')]
    temp.data$pigeon = dimnames(data)[[3]][j]
    
    # support and cross wind
    speed = get_dist(data[,"lon",j],
                     data[,"lat",j], hz = 5, method = "speed")
    temp.data$speed = speed
    if ( i > 8){
      as.s = air_speed(data[,1:2,j],
                       wind.direct = md$wind.direct.site,
                       wind.speed  = md$wind.speed.site ,
                       speed, 
                       return.support.cross = T)
      temp.data$support.wind =  as.s[[2]]
      temp.data$cross.wind   =  as.s[[3]]
    }
    
    # focal heading and turning decisions. autocorrelation
    focal_head = get_heading(data[,"lon",j],data[,"lat",j])
    temp.data$focal_head = focal_head
    temp.data$diff_head = turnang(data[,,j])
    
    # lon lat
    temp.data$lon = data[,"lon",j]
    temp.data$lat = data[,"lat",j]
    
    # dist 2 cent
    temp.data$dist2cent = d2cent[,j]
    
    # turn angle to home
    ang2home = get_heading( data[,"lon",j],
                            data[,"lat",j],
                            home$lon,
                            home$lat,indivs = 2)
    temp.data$turn2home = atan2(sin(ang2home-focal_head), cos(ang2home-focal_head))
    
    # turn angle to predator position, turn to pred head
    if ( condition[i] == "p"){
      
      # falcon's relative position
      ang2falcpos = get_heading( data[,"lon",j],
                                 data[,"lat",j],
                                 falc[,"lon"],
                                 falc[,"lat"],indivs = 2)
      temp.data$turn2falcpos = atan2(sin(ang2falcpos-focal_head),
                                     cos(ang2falcpos-focal_head))
      # falcon's relative heading
      falc_head = get_heading(falc[,"lon"],falc[,"lat"])
      temp.data$turn2falchead = atan2(sin(falc_head-focal_head),
                                      cos(falc_head-focal_head))
      # distance to falcon
      temp.data$dist2pred = get_dist(data[,"lon",j],data[,"lat",j],falc[,"lon"],falc[,"lat"], method = "distance")
      temp.data$pred.lat      = falc[,"lat"]
temp.data$pred.lon      = falc[,"lon"]
      
    } else {
      temp.data$turn2falcpos  = NA
      temp.data$turn2falchead = NA
      temp.data$dist2pred     = NA
      temp.data$pred.lat      = NA
      temp.data$pred.lon      = NA
    }
    
    # distance to home 
    temp.data$dist2home = get_dist(data[,"lon",j],data[,"lat",j],home$lon,home$lat,method = "distance")
  }
    # near neighbours 
    nn.dist = nn[[1]] 
    nn.name = nn[[3]]
    cent.head = nns.centroid.heading( data = data,
                                      heads = heads,
                                      data.remain= data.remain,
                                      heads.remain = heads.remain,
                                      nn.name = nn.name,
                                      j = j,
                                      num.nei = 33)
    
    temp.data = cbind( temp.data, cent.head)
    temp.data$nn1dist = nn[[2]][,j]
    
    # dist2site
    temp.data$dist2site =  get_dist(data[,"lon",j],data[,"lat",j],centroid[1,"lon"],centroid[1,"lat"], method = "distance")
    
    # big.matrix
    temp.data$unique.flight = m
    temp.data$study.flight = i
    li[[m]] = temp.data
    m = m+1
  }
  
  # takestock
  print(paste ( i, "/" , length(files)))
}

# convert to dataframe
{
  dat = ldply (li)
  dat.num = dat
  for ( i in 1:ncol(dat)){
    dat.num[,i] = suppressWarnings(as.numeric(as.character(dat[,i])))
    if ( length(which(!is.na(dat.num[,i]))) == 0){
      dat.num[,i] = as.character(dat[,i])
    }
  }
  dat = dat.num
}

# save

if ( only.remaining.indivs){
  save( dat, file = file.path(PROJHOME , "gitData", "REMrobofalcon-data.rda" ))
} else {
  save( dat, file = file.path(PROJHOME, "gitData", "ALLrobofalcon-data.rda" ))
}

# take a look
head(dat)
```


# time lags 

Run this script with only.remaining.indivs = T and = F

```{R}
# housekeeping
rm(list = ls())

# objects
only.remaining.indivs = T

# files/folders
if( only.remaining.indivs ){
  load( file.path(PROJHOME , "gitData", "REMrobofalcon-data.rda" ))
} else {
  load( file.path ( PROJHOME , "gitData", "ALLrobofalcon-data.rda"))
}

# each unique bird/flight ends here.
li = list()
m=1
fli = c( 0, which(diff(dat$unique.flight) == 1),nrow(dat))
lags = seq(-5,2,1)

for ( i  in 1:(length(fli)-1)){
  #i = 1
  d = dat[(fli[i]+1):fli[i+1],]
  lagmat = matrix(NA,nrow(d),ncol=length(lags),
                  dimnames = list(NULL,paste0("lag",lags)))
  lagmat[,"lag0"] = d$diff_head
  
  for ( j in 1:length(lags) ){
    # j = 3
    k = lags[j]
    if ( sign(k) == -1){
      lagmat[,j] =  c(d$diff_head[(-k + 1):nrow(d)],rep(NA,-k))
    }
    if ( sign(k) == 1 ){
      lagmat[,j] = c(rep(NA,k),d$diff_head[1:(nrow(d)-k)])
    }
  }
  # save
  d = cbind( d, lagmat)
  li[[i]] = d
}


# convert to dataframe
{
  dat = ldply (li)
  dat.num = dat
  for ( i in 1:ncol(dat)){
    dat.num[,i] = suppressWarnings(as.numeric(as.character(dat[,i])))
    if ( length(which(!is.na(dat.num[,i]))) == 0){
      dat.num[,i] = as.character(dat[,i])
    }
  }
  dat = dat.num
}

#save
if ( only.remaining.indivs){
  save( dat, file = file.path(PROJHOME , "gitData", "REMrobofalcon-data.rda" ))
} else {
  save( dat, file = file.path(PROJHOME, "gitData", "ALLrobofalcon-data.rda" ))
}

# take a look
head(dat)
```


# Proportion separated

```{R}
{ # skip to loop
  
  # housekeeping
  rm(list = ls())
  
  # libraries
  library(robofalcon)
  library(stringr)
  library(lme4)
  library(nlme)
  
  # files / folders 
  fold = file.path(PROJHOME, "Output", "dataremovedoutliers")
  files = list.files(fold)
  spl = str_split_fixed(files, "\\.",12)
  
}

fission.data = matrix( NA, 0, 4, dimnames =  list( NULL, c("flock.size", "condition" , "id" , "fission")))
# loop
k = 1
for ( i in 9:length(files)){
  #i=9
  # data
  load( file.path( fold, files[i]))
  
  data = data.list[[1]]
  dims = dim(data)
  names = dimnames(data)[[3]]
  
  for (j in 1:dims[3]){
    #j=4
    fission.data = rbind( fission.data, rep(NA,ncol(fission.data)))
    fission.data[k,"fission"] = length( which ( is.na ( data[,"lat",j])))/dims[1]
    fission.data[k,"id"] = names[j]
    fission.data[k,"flock.size"] = spl[i,7]
    fission.data[k,"condition"] = spl[i,2]
    k = k+1
  }
}
fission.df = as.data.frame(fission.data)
fission.df$fission = as.numeric(as.character(fission.df$fission))


save ( fission.df , file = file.path (PROJHOME , "smaller-dataframes-for-git", "fission.rda"))
```

# Descriptives. Exact numbers of each pigeon, condition, 
```{R}
# housekeeping
rm(list = ls())

# folders/ files
fold = file.path ( PROJHOME , "Data" , "Pigeon")
files = list.files ( fold)

# matrix of flights
mat = stringr::str_split_fixed(files, "\\." , 12)
mat = as.data.frame(mat)
names(mat) = c("unique.flight", "pigeon", "condition", "groupID", "flight.num", "small.big", "day", "month", "gps", "acc1", "acc2", "filetype")
levels(mat$condition) = droplevels(mat$condition)
tb = table(mat$pigeon, mat$groupID)[,1:4]
group.id = apply(tb, 1, which.max )

# table
cb = cbind( t(table( mat$condition,mat$pigeon, mat$small.big)[1:2,,1]),
            t(table( mat$condition,mat$pigeon, mat$small.big)[1:2,,2]))
cb = as.data.frame(cb)
cb$id = rownames(cb)
cb$group.id = group.id

# save
folderDS( c("Output", "descriptives"))
write.csv(cb, file = file.path(PROJHOME , "Output" , "descriptives" , "eachPidge_flightsperCondition_table.csv"))
```
