==> Reports/03_conflict-scenario.Rmd <==
---
title: "R Notebook"
output: html_notebook
---



# Conflict scenarios and other last variables before analysis 

```{R}

# housekeeping
rm(list = ls())

library( robofalcon)

# data
{
  load( file.path( PROJHOME , "Output" , "Topological3", "REMrobofalcon-data.rda"))
  dir.rel.ali = dat$align - dat$diff_head
  dat$sig.cent = sign( dat$futatt)
  dat$sig.head = sign( dat$align)
  dat$conflict = ifelse( dat$sig.head != dat$sig.cent, 1,-1) # binary conflict scenario variable. Did you turn toward centroid or away from predator
  dat$dir.rel.ali = dir.rel.ali * dat$conflict # Instanstaneous centroid attraction
  
  dat$sig.pred = sign( dat$turn2falchead)
  dat$conflict.pred = ifelse( dat$sig.head != dat$sig.pred, 1,-1)
  dat$dir.rel.pred = dir.rel.ali * dat$conflict.pred
  dat$sign.diff = sign(dat$diff_head)
  dat$sqrt_diff_head = dat$sign.diff * sqrt(abs(dat$diff_head))
  dat$pigeon = as.character(dat$pigeon)
  dat$group.num = as.character(dat$group.num)
  dat$unique.flight = as.character(dat$unique.flight)
  dat.coh = dat
}

{
  load( file.path( PROJHOME , "Output" , "Topological", "ALLrobofalcon-data.rda"))
  dir.rel.ali = dat$align - dat$diff_head
  dat$sig.cent = sign( dat$futatt)
  dat$sig.head = sign( dat$align)
  dat$conflict = ifelse( dat$sig.head != dat$sig.cent, 1,-1)
  dat$dir.rel.ali = dir.rel.ali * dat$conflict
  
  dat$sig.pred = sign( dat$turn2falchead)
  dat$conflict.pred = ifelse( dat$sig.head != dat$sig.pred, 1,-1)
  dat$dir.rel.pred = dir.rel.ali * dat$conflict.pred
  dat$sign.diff = sign(dat$diff_head)
  dat$sqrt_diff_head = dat$sign.diff * sqrt(abs(dat$diff_head))
  dat$pigeon = as.character(dat$pigeon)
  dat$group.num = as.character(dat$group.num)
  dat$unique.flight = as.character(dat$unique.flight)
  dat$speed = get_dist(dat$lat,dat$lon, method = "speed" , hz = 5)
  dat$turn.radius = dat$speed / abs(dat$diff_head)
  dat$centrip = dat$speed^2 / dat$turn.radius # Centripetal acceleration
  dat.coh$centrip = dat$centrip
}


dat$sign.diff = sign(dat$diff_head)
dat$diff_head[ abs(dat$diff_head) > 0.3] = NA # See supplemental for validation
dat$dir.rel.ali[ abs(dat$dir.rel.ali)  > 0.3] = NA
dat$diff_head_sharp = dat$diff_head
dat$dir.rel.ali_sharp = dat$dir.rel.ali
dat$diff_head_sharp[ abs(dat$diff_head_sharp) < 0.02] = NA # See supplemental for validation
dat$dir.rel.ali_sharp[ abs(dat$dir.rel.ali_sharp) < 0.02] = NA
dat.coh$diff_head_sharp = dat$diff_head_sharp
dat$sqrt_diff_head = dat$sign.diff * sqrt(abs(dat$diff_head))


# CONFLICT
dat$sig.cent = sign( dat$futatt)
dat$sig.ali  = sign( dat$align)
dat$sig.head = sign( -dat$turn2falchead)
dat$sig.turn = sign( dat$diff_head)

# only consider the half of the group closest to the falcon
source ( file.path  (PROJHOME , "source" , "onlyHalfInConflict.r"))

dat$conflict.cent = ( dat$sig.head != dat$sig.cent)
con = which(dat$conflict.cent == T)
dat$biased.cent = dat$sig.turn == dat$sig.cent
dat$conflict.ali = ( dat$sig.head != dat$sig.ali)
con = which(dat$conflict.ali == T)
dat$biased.ali = dat$sig.turn == dat$sig.ali
dat.con = dat[con,]
dat.nocon = dat[-con,]
dat.con = dat.con  [!is.na ( dat.con$dist2predNearest),]
dat.nocon=dat.nocon[!is.na(dat.nocon$dist2predNearest),]

length ( which ( dat$conflict.ali)) / nrow(dat)

# save 
folderDS ( c ( "Output" , "FinalDataSet"))
folderDS ( c ( "Output" , "ConflictScen"))
save ( dat    , file = file.path ( PROJHOME , "Output", "FinalDataSet", "ALLrobofalcon-data.rda"))
save ( dat.coh, file = file.path ( PROJHOME , "Output", "FinalDataSet", "REMrobofalcon-data.rda"))
save ( dat.con  , file = file.path ( PROJHOME , "Output", "ConflictScen", "DatCon.rda"          ))
save ( dat.nocon, file = file.path ( PROJHOME , "Output", "ConflictScen", "DatNoCon.rda"        ))
```
