==> Reports/04_models.rmd <==
# All Models

# What impacts turning angle deltaT? (or diff_head as it is called in these models) It is called diff_head because it is essentially the difference in heading from last timestep to the present angle.



```{R}
library(robofalcon)
# make some new folders to save model data
folderDS( c("Output", "Statistics" , "Tables")  )
folderDS( c("Output", "Statistics" , "Models")  ) 
```

# Model 1. THIS IS NOT USED AS ONE OF THE FINAL MODELS DUE TO LACK OF COHERENCE WITH PARAMETRIC ASSUMPTIONS. All recorded instances of diff_head, including turning and straight line flight. These results are reported in the supplemental. Table S4 in current version, labelled as "Turn angle - all data"

```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
#library(dplyr)
#library(EMAtools)
library(robofalcon)

# LOAD DATA and other variables
source( file.path(PROJHOME , "source", "source.r"))

# SUBSET DATA
d = dat.coh[sam9,] # subsample every 9 datapoints to correct for autocorrelation. See supplemental material. 

# REMOVE VARIABLES NOT OF INTEREST TO THE MODEL
d = d[,c("diff_head", "condition","futatt", "align", "cross.wind" , "turn2home", "pigeon", "unique.flight", "study.flight", "group.num", "binary.closest")]
d = d[complete.cases(d),]

# RUN MODEL
lmm0 = lme( diff_head ~
              condition *futatt +
              condition *align  +
              cross.wind +
              turn2home
            , random = list( pigeon=~1,
                             study.flight =~1,
                             group.num = ~1
                             # ,unique.flight = ~1
                             ), data = d, na.action = "na.fail" )

# DIAGNOSTICS
diagnostics.plot(lmm0 )
summary(lmm0)

# SAVE
save.mixed.model( lmm0 , d, "00_Turn-angle-all.csv") # look at this function in folder /R to see all output statistics, or refer to supplemental Table S4 where they are all pasted. 
```

The qq plot and residual distributions are off. This is due to the inclusion of straight line flight. See supplemental and the below model for correction


# Model 1.1. Turning angles after removing straight line flight. See supplemental for validation of this approach. These results are reported in the main text



```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
#library(EMAtools)

# variables

# LOAD DATA and other variables
source( file.path(PROJHOME , "source", "source.r"))

# SUBSET DATA

d.coh = dat.coh[sam9,]
d.coh = d.coh[,names(d.coh) %in% c("diff_head_sharp","condition","futatt", "study.flight", "align", "cross.wind" , "turn2home", "pigeon", "unique.flight", "group.num", "small.big")]
d.coh = d.coh[complete.cases(d.coh),]

# MODEL
lmm1 = lme( diff_head_sharp ~
               condition *
              futatt+
               condition *
              align +
                cross.wind +
                turn2home
              , random = list( pigeon=~1,
                               study.flight =~1,
                               group.num = ~1
                               #, unique.flight = ~1
                               ), data = d.coh, na.action = "na.fail" )

# DIAGNOSTICS
diagnostics.plot(lmm1)
summary(lmm1)

# SAVE
save.mixed.model( lmm1 ,d.coh, "01_Turn-angle-sharp.csv")
```



# Model 1.2. remove interaction terms. 



```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
#library(EMAtools)



# LOAD DATA and other variables
source( file.path(PROJHOME , "source", "source.r"))

# SUBSET DATA

d.coh = dat.coh[sam9,]
d.coh = d.coh[,names(d.coh) %in% c("diff_head_sharp","condition","futatt", "study.flight", "align", "cross.wind" , "turn2home", "pigeon", "unique.flight", "group.num", "small.big")]
d.coh = d.coh[complete.cases(d.coh),]

# MODEL
lmm1 = lme( diff_head_sharp ~
               # condition *
              futatt+
               # condition *
              align +
                cross.wind +
                turn2home
              , random = list( pigeon=~1,
                               study.flight =~1,
                               group.num = ~1
                               #, unique.flight = ~1
                               ), data = d.coh, na.action = "na.fail" )

# DIAGNOSTICS
diagnostics.plot(lmm1)
summary(lmm1)

# SAVE
save.mixed.model( lmm1 ,d.coh, "01_2_Turn-angle-sharp.csv")

d.coh = d.coh[d.coh$condition == "c",]
ggplot( d.coh, aes( y = diff_head_sharp, x = futatt))+
  geom_smooth()+
  xlim(c(-2,2))
```
### Model 2 - remove data from when predator was close

```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
#library(EMAtools)
library(robofalcon)

# LOAD DATA and other variables
source( file.path(PROJHOME , "source", "source.r"))

# variables 
collision.dist = 0.8

# SUBSET DATA

d.coh = dat.coh[ is.na(dat.coh$dist2pred ) | dat.coh$dist2pred >40,]
nrow( d.coh)/ nrow(dat.coh)
d.coh = d.coh[sam9,]
d.coh = d.coh[which ( d.coh$minAlignNei > collision.dist & d.coh$minTowardCent > collision.dist),]
d.coh = d.coh[,names(d.coh) %in% c("diff_head_sharp","condition","futatt", "study.flight", "align", "cross.wind" , "turn2home", "pigeon", "unique.flight", "group.num", "small.big")]
d.coh = d.coh[complete.cases(d.coh),]


# MODEL
lmm2 = lme( diff_head_sharp ~
                condition *futatt+
                condition *align +
                cross.wind +
                turn2home
              , random = list( pigeon=~1,
                               study.flight =~1,
                               group.num = ~1
                               # ,unique.flight = ~1
                               ), data = d.coh, na.action = "na.fail" )

# DIAGNOSTICS
diagnostics.plot(lmm2)
summary(lmm2)

# SAVE
save.mixed.model( lmm2 ,d.coh, "02_Turn-angle-coh.csv")


```

# MODEL 3. Differences between small and large flocks? 




```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
library(EMAtools)
library(robofalcon)

# LOAD DATA and other variables
source( file.path(PROJHOME , "source", "source.r"))

# SUBSET - ALL VARIABLES - so now including FALCON
d.coh = dat.coh#
#d.coh = d.coh [d.coh$binary.closest==T,]
#d.coh = d.coh[which ( d.coh$minAlignNei > collision.dist & d.coh$minTowardCent > collision.dist),]
d = d.coh[sam9,]
d = d[,c("turn2falchead" , "turn2falcpos", "diff_head_sharp","futatt", "study.flight", "align", "cross.wind" , "turn2home", "pigeon", "unique.flight", "group.num", "small.big", "binary.closest", "dist2pred")]
d = d[complete.cases(d),]

# MODEL
lmm3 = lme( diff_head_sharp ~
              futatt * small.big 
          *binary.closest
            +
              align  * small.big 
          *binary.closest
            +
              turn2falchead +
              turn2falcpos +
              cross.wind +
              turn2home
            , random = list( pigeon=~1,
                             group.num = ~1,
                             study.flight=~1), data = d,
            na.action = "na.fail" )

# DIAGNOSTICS
  diagnostics.plot(lmm3)
summary(lmm3)

# SAVE
mod = save.mixed.model( lmm3 , d, "03_Turn-angle-falc.rda")
mod

ggplot( d, aes( y = diff_head_sharp, x = futatt))+
  geom_smooth(aes(colour = factor(binary.closest)))+
  xlim(c(-1,1))
```

```{R}
d.T = d[ d$binary.closest == T,]

lmm3 = lme( diff_head_sharp ~
              futatt * small.big  +
              align  * small.big   +
              turn2falchead +
              turn2falcpos +
              cross.wind +
              turn2home
            , random = list( pigeon=~1,
                             group.num = ~1,
                             study.flight=~1), data = d.T,
            na.action = "na.fail" )
mod = save.mixed.model(lmm3, d.T,"Just_Further.rda")
mod

```

```{R}

d.F = d[ d$binary.closest == F,]

lmm3 = lme( diff_head_sharp ~
              futatt * small.big  +
              align  * small.big   +
              turn2falchead +
              turn2falcpos +
              cross.wind +
              turn2home
            , random = list( pigeon=~1,
                             group.num = ~1,
                             study.flight=~1), data = d.F,
            na.action = "na.fail" )
mod = save.mixed.model(lmm3, d.F,"Just_Further.rda")
mod



```

```{R}
d.40 = d[ is.na(d$dist2pred ) | d$dist2pred >40,]

lmm3 = lme( diff_head_sharp ~
              futatt * small.big 
          *binary.closest
            +
              align  * small.big 
          *binary.closest
            +
              turn2falchead +
              turn2falcpos +
              cross.wind +
              turn2home
            , random = list( pigeon=~1,
                             group.num = ~1,
                             study.flight=~1), data = d.40,
            na.action = "na.fail" )

summary(lmm3)
mod =  save.mixed.model( lmm3 , d.40, "03_2_Turn-angle-falc.csv")
ggplot( d.40, aes( y = diff_head_sharp, x = futatt))+
  geom_smooth(aes(colour = factor(binary.closest)))+
  xlim(c(-1,1))
mod

```


```{R}
d.40 = d[ is.na(d$dist2pred ) | d$dist2pred >40,]

lmm3 = lme( diff_head_sharp ~
              futatt * small.big 
          *binary.closest
            +
              align  * small.big 
          *binary.closest
            +
              turn2falchead +
              turn2falcpos +
              cross.wind +
              turn2home
            , random = list( pigeon=~1,
                             group.num = ~1,
                             study.flight=~1), data = d.40,
            na.action = "na.fail" )

summary(lmm3)
mod =  save.mixed.model( lmm3 , d.40, "03_2_Turn-angle-falc.csv")
ggplot( d.40, aes( y = diff_head_sharp, x = futatt))+
  geom_smooth(aes(colour = factor(binary.closest)))+
  xlim(c(-1,1))

```



# MODEL 4. Speed variance

```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
library(EMAtools)

# LOAD DATA and other variables
source( file.path(PROJHOME , "source", "source.r"))
load(file.path(PROJHOME,"smaller-dataframes-for-git" , "Speedvardat.rda"))
dat$speed.var = speed.vars

# SUBSET
da = dat[sam16,]
d =  da[,c("speed.var", "condition","small.big", "speed", "cross.wind","dist2site", "support.wind","unique.flight","group.num")]

# Transform - see Supplemental
b = MASS::boxcox((d$speed.var)  ~  condition *  small.big +
                   abs( cross.wind) +
                   support.wind, data = d)
lamda=b$x
lik=b$y
bc = cbind(lamda, lik)
head(bc[order(-lik),])
cb =bc[order(-lik),]
d$trans.speed.var = d$speed.var^cb[2,1] 

# model
d = d[complete.cases(d),]
lmm4 = lme( trans.speed.var ~
              condition *  small.big +
              abs( cross.wind) +
              support.wind
            , random = list( group.num=~1,
                             unique.flight =~1),
            data = d, na.action = "na.fail" )

# summary table
summary(lmm4)

# save
save.mixed.model(lmm4, d,"04_Speedvariance.csv")
```

# MODEL 5. CENTRIPETAL ACCELERATION

```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
library(EMAtools)

# LOAD DATA 
source( file.path(PROJHOME , "source", "source.r"))

# subsample
d = dat[sam6,]
d = d[,c("condition", "centrip", "small.big","cross.wind", "support.wind", "group.num", "pigeon", "unique.flight")]

# cut off - See supplemental
d$centrip [d$centrip < 0.01] = NA
d$centrip [d$centrip > 13 ] = NA

# remove NA
d = d[complete.cases(d),]

# transform
b = MASS::boxcox((d$centrip)  ~  condition * small.big +
                   abs(cross.wind) +
                   support.wind , data = d)
lamda=b$x
lik=b$y
bc = cbind(lamda, lik)
head(bc[order(-lik),])
cb =bc[order(-lik),]
d$trans.centrip = d$centrip^cb[1,1]

# MODEL
lmm5 = lme( trans.centrip ~
              condition * small.big +
              abs(cross.wind) +
              support.wind 
            , random = list( group.num=~1,
                             pigeon=~1,
                             unique.flight =~1), data = d,
            na.action = "na.fail" )

# summary table
summary(lmm5)

# Save
save.mixed.model(lmm5, d,"05_centrip.csv")
```

# MODEL 6. SPEED
```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
library(EMAtools)

# LOAD DATA 
source( file.path(PROJHOME , "source", "source.r"))

# SUBSAMPLE
d = dat.coh[sam24,]
d = d[,c("condition", "speed", "small.big","cross.wind", "support.wind", "group.num", "pigeon", "unique.flight")]
d$speed[d$speed> 35] = NA # See supplemental
d$speed[d$speed< 5]  = NA
d = d[complete.cases(d),]

# MODEL
lmm6 = lme( speed ~
              condition * small.big +
              abs(cross.wind) +
              support.wind, 
            random = list( 
              group.num=~1,
              pigeon=~1,
              unique.flight =~1), data = d,
            na.action = "na.fail" )

# Summary table
summary(lmm6)

# Save
save.mixed.model(lmm6, d,"06_Speed.csv")
```

# MODEL 7. FLOCK COHESION

```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
library(EMAtools)

# LOAD DATA
source( file.path(PROJHOME , "source", "source.r"))
load( file.path(PROJHOME , "smaller-dataframes-for-git", "insta-grp-size.rda"))
dat.coh$insta.grp.size = insta.grp.size

# SUBSAMPLE
d = dat.coh[sam27,]
d = d[,c( "insta.grp.size", "dist2cent", "condition", "cross.wind", "support.wind", "group.num", "unique.flight", "pigeon")]
d = d[complete.cases(d),]

# MODEL
lmm7 = lme( log(dist2cent+1) ~
              insta.grp.size * condition +
              abs(cross.wind)+
              support.wind
            , random = list( group.num=~1,unique.flight=~1, pigeon=~1), data = d, na.action = "na.fail" )

# Summary table
summary(lmm7)

# Save
save.mixed.model(lmm7,d,"07_dist2cent-vs-condition.csv")
```

# MODEL 8. NEAREST NEIGHBOUR DISTANCE

```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
library(EMAtools)
library(robofalcon)

# LOAD DATA
source( file.path(PROJHOME , "source", "source.r"))
load( file.path(PROJHOME , "smaller-dataframes-for-git", "insta-grp-size.rda"))
dat.coh$insta.grp.size = insta.grp.size

# SUBSAMPLE
d = dat.coh[sam26,]
d = d[,c( "insta.grp.size", "nn1dist", "condition", "cross.wind", "support.wind", "group.num", "unique.flight", "pigeon")]
d$nn1dist[d$nn1dist > 10] = NA # See supplemental for validation
d = d[complete.cases(d),]

# Transform
b = MASS::boxcox((d$nn1dist+1)  ~ insta.grp.size * condition +
                   abs(cross.wind) +
                   support.wind , data = d)
lamda=b$x
lik=b$y
bc = cbind(lamda, lik)
head(bc[order(-lik),])
cb =bc[order(-lik),]
d$trans.nn1 = d$nn1dist^cb[1,1]
#d$trans.nn1 = as.numeric(as.character(d$nn1dist^0.25))
d$trans.nn1[is.infinite(d$trans.nn1)] = NA
d = d[complete.cases(d),]

# MODEL
lmm8 = lme( log( trans.nn1+1)~
              insta.grp.size * condition  +
              abs(cross.wind) +
              support.wind 
            , random = list( group.num=~1, unique.flight=~1, pigeon = ~1), data = d, na.action = "na.omit" )

# Summary table
summary(lmm8)

# Save
save.mixed.model(lmm8,d,"08_nn1-vs-condition.csv")
```

# MODEL 9. Proportion separated

```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
library(EMAtools)
library(robofalcon)

# LOAD DATA 
load( file.path(PROJHOME , "smaller-dataframes-for-git", "fission.rda"))
fission.df$prop.separated = fission.df$fission # Fission means something else in this paper. Please ignore the use of the term here

# MODEL
lmm9 = lme( prop.separated ~
              condition * flock.size 
            , random = list( id=~1)
            , data = fission.df)

# Summary table
summary(lmm9)

# Save
save.mixed.model (lmm9,fission.df, "09_proportion separated.csv")
```

# MODEL 10. Conflict scenarios

```{R}
#housekeeping
rm(list = ls())


# libraries
library(ggplot2)
library(nlme)
library(EMAtools)

# LOAD DATA 
source( file.path(PROJHOME , "source", "source.r"))
#load( file.path ( PROJHOME , "Output", "ConflictScen", "DatCon.rda"))

# Autocorrelation - different to other autocorrelation, as timesteps are already staggered. See Supplemental.
dat.con =  dat[ dat$conflict.cent == T,]
ts.acf <- acf(as.numeric(na.omit(dat.con$biased.cent)))
alpha = 0.95
conf.lim <- qnorm((1 + alpha)/2)/sqrt(ts.acf$n.used)
aep = which(ts.acf$acf < conf.lim)[1]


# SUBSAMPLE
sam = seq(1,nrow(dat.con),aep)
d.con = dat.con[sam,c("biased.cent","dist2pred","small.big","pigeon","unique.flight", "group.num","study.flight")]
d.con = d.con[complete.cases(d.con),]

# MODEL
mod10 = MASS::glmmPQL( biased.cent~
                       log(dist2pred)* small.big
                      , random = list( pigeon=~1,unique.flight =~1,
                                       group.num = ~1
                                       #, study.flight = ~1
                                       ),
                      family = "binomial", data = d.con,
                      na.action = "na.fail" )

# Summary table
summary(mod10)

# Save
save.mixed.model(mod10 , d.con, "10_conflict.csv")
```



# MODEL 11. No conflict

```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
library(EMAtools)

# LOAD DATA 
source( file.path(PROJHOME , "source", "source.r"))
load( file.path ( PROJHOME , "Output", "ConflictScen", "DatNoCon.rda"))

# Autocorrelation
ts.acf <- acf(as.numeric(na.omit(dat.nocon$biased.cent)))
alpha = 0.95
conf.lim <- qnorm((1 + alpha)/2)/sqrt(ts.acf$n.used)
aep = which(ts.acf$acf < conf.lim)[1]
sam = seq(1,nrow(dat.nocon),aep)

# SUBSET
d.con = dat.nocon[sam,c("biased.cent","dist2pred","small.big","pigeon","unique.flight", "group.num")]
d.con = d.con[complete.cases(d.con),]

# MODEL
mod11 = MASS::glmmPQL( biased.cent~
                         dist2pred * small.big
                       , random = list( pigeon=~1,unique.flight =~1,
                                        group.num = ~1),
                       family = "binomial", data = d.con,
                       na.action = "na.fail" )

# Summary table
summary(mod11)

# Save
save.mixed.model(mod11 , d.con, "11_noconflict.csv")
```

# MODEL 12,13,14. Fission/fusion at different sensitivity parameters, 1 second 5Hz, 2 seconds 10Hz and 3 seconds 15Hz 

```{R}
# Housekeeping
rm(list = ls())

## libraries
library(ggplot2)
library(nlme)
library(EMAtools)

#LOAD DATA. Make biomial 
load( file.path(PROJHOME, "smaller-dataframes-for-git", "splitdata-minsplit-5Hz.rda" ))
sd = sd[ -which (sd$splitornot == "nosplit"),]
  sd = sd[ sd$condition == "p",]
  sd = sd[(complete.cases(sd[,c("dist2pred")])),]
  sd = sd[ !is.na(sd$splitornot),]
  sd$binom = as.numeric(as.factor(sd$splitornot))-1 # this makes rejoin/fusion = 0 and split/fission = 1. See Fig. 4 in main text
dat5 = sd
load( file.path(PROJHOME, "smaller-dataframes-for-git", "splitdata-minsplit-10Hz.rda" ))
sd = sd[ -which (sd$splitornot == "nosplit"),]
  sd = sd[ sd$condition == "p",]
  sd = sd[(complete.cases(sd[,c("dist2pred")])),]
  sd = sd[ !is.na(sd$splitornot),]
  sd$binom = as.numeric(as.factor(sd$splitornot))-1
dat10 = sd
load( file.path(PROJHOME, "smaller-dataframes-for-git", "splitdata-minsplit-15Hz.rda" ))
sd = sd[ -which (sd$splitornot == "nosplit"),]
  sd = sd[ sd$condition == "p",]
  sd = sd[(complete.cases(sd[,c("dist2pred")])),]
  sd = sd[ !is.na(sd$splitornot),]
  sd$binom = as.numeric(as.factor(sd$splitornot))-1
dat15 = sd

# MODELS
mod12 = MASS::glmmPQL( binom ~ log(dist2pred)*small.big      , random = list(pigeon=~1,group.num=~1) , family = "binomial" , data = dat5, na.action = na.omit)
mod13 = MASS::glmmPQL( binom ~ log(dist2pred)*small.big      , random = list(pigeon=~1,group.num=~1) , family = "binomial" , data = dat10, na.action = na.omit)
mod14 = MASS::glmmPQL( binom ~ log(dist2pred)*small.big      , random = list(pigeon=~1,group.num=~1) , family = "binomial" , data = dat15, na.action = na.omit)

# summary tables
summary(mod12)
summary(mod13)
summary(mod14)

# save
save.mixed.model(mod12, dat5 ,"12_fissionFusion5Hz.csv")
save.mixed.model(mod13, dat10,"13_fissionFusion10Hz.csv")
save.mixed.model(mod14, dat15,"14_fissionFusion15Hz.csv")
```



# MODEL 15. Do pigeons differ in their selfishness? Seeing if selfish herd parameter (dir.rel.ali) is a better model with random intercepts or not will tell us. If the model with random intercepts for each pigeon is a better model, they might be more/less selfish than one another. If the AICs are similar, they are using the selfish herd parameter to a similar extent. See discussion in supplemental. 


```{R}
#housekeeping
rm(list = ls())

# libraries
library(ggplot2)
library(nlme)
library(EMAtools)

# LOAD DATA 
source( file.path(PROJHOME , "source", "source.r"))

d = dat[sam9, ]
lmm15 = lme( dir.rel.ali_sharp~
               condition * dist2cent 
             , random = list( unique.flight=~1, group.num = ~1, pigeon = ~1), data = d, na.action = "na.omit" )

lmm15.1 = lme( dir.rel.ali_sharp~
                 condition * dist2cent
               , random = list( unique.flight=~1, group.num = ~1), data = d, na.action = "na.omit" )
AIC(lmm15)
AIC(lmm15.1)
```

```{R}
summary(lmm15)
```

This model also shows that predator groups use more centroid attraction at higher distance to centroid. Similar to the three way interaction reported in Results (main text)

# Combine all statistics into a big table

```{R}
########### BIG TABLE

rm(list = ls())
fold = file.path ( PROJHOME , "Output" , "Statistics" , "Tables")
files = rev(list.files(fold))
d = read.csv( file.path(fold, files[1]))
d$name = files[1]
dat = d
for ( i in 1:length(files)){
  #i=1
  d = read.csv( file.path(fold, files[i]))
  d$name = files[i]
  dat = rbind( d, dat)
}

numer = which ( unlist(lapply( dat,class)) == "numeric")
for ( i in numer){
  dat[,i] = round(as.numeric(as.character(dat[,i])),3)
}

folderDS(c("Output", "Statistics"))
write.csv(dat , file = file.path ( PROJHOME, "Output", "Statistics" , "combinedstatstables.csv"))
```
